Room management >edit page

+++ StudioReservation.BackOffice/Controllers/RoomController.cs
@@ -2,11 +2,8 @@
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
-using System.Runtime.InteropServices;
-using System.Runtime.InteropServices.ComTypes;
 using System.Web;
 using System.Web.Mvc;
-using System.Web.Security;
 using StudioMiscellaneous.Service.Contract;
 
 //using StudioReservation.DataModel;
@@ -25,59 +22,6 @@ namespace StudioReservation.BackOffice.Controllers
         public ActionResult Index()
 
         {
-            //var result = _reservationService.FindAllRoomTimeSlot();
-            //RoomsViewModel roomViewModel = new RoomsViewModel();
-            //List<Room> roomList = new List<Room>()
-            //{
-            //    new Room {
-            //        Id =  1,
-            //        Name = "Studio 1",
-            //        Description = "Measuring over 800 sqft and can accommodate up to 16 dancers comfortably. Our dance floor rests on a system of high-density rubber underlays and laminate wood flooring. It is fully air-conditioned and equipped with barres, mirrors and sound system compatible for audio input jack, USB device and SD card. This studio overlooks the street with many windows for natural lighting and ventilation.",
-            //        Image = "Studio1.jpg",
-            //        Size = "800 * 750",
-            //        Style= "K-pop;Hip Hop;Jazz",
-            //        Rate = 100.00M
-            //      },
-            //      new Room {
-            //        Id =  2,
-            //        Name = "Studio 2",
-            //        Description = "Measuring over 800 sqft and can accommodate up to 16 dancers comfortably. Our dance floor rests on a system of high-density rubber underlays and laminate wood flooring. It is fully air-conditioned and equipped with barres, mirrors and sound system compatible for audio input jack, USB device and SD card. This studio overlooks the street with many windows for natural lighting and ventilation.",
-            //        Image = "Studio2.jpg",
-            //        Size = "800 * 750",
-            //        Style= "K-pop;Hip Hop;Jazz",
-            //        Rate = 120.00M
-            //      },
-            //      new Room {
-            //        Id =  3,
-            //        Name = "Studio 3",
-            //        Description = "Measuring over 800 sqft and can accommodate up to 16 dancers comfortably. Our dance floor rests on a system of high-density rubber underlays and laminate wood flooring. It is fully air-conditioned and equipped with barres, mirrors and sound system compatible for audio input jack, USB device and SD card. This studio overlooks the street with many windows for natural lighting and ventilation.",
-            //        Image = "Studio3.jpg",
-            //        Size = "800 * 750",
-            //        Style= "K-pop;Hip Hop;Jazz",
-            //        Rate = 100.00M
-            //      },
-            //      new Room {
-            //        Id =  4,
-            //        Name = "Studio 4",
-            //        Description = "Measuring over 800 sqft and can accommodate up to 16 dancers comfortably. Our dance floor rests on a system of high-density rubber underlays and laminate wood flooring. It is fully air-conditioned and equipped with barres, mirrors and sound system compatible for audio input jack, USB device and SD card. This studio overlooks the street with many windows for natural lighting and ventilation.",
-            //        Image = "Studio4.jpg",
-            //        Size = "800 * 750",
-            //        Style= "K-pop;Hip Hop;Jazz",
-            //        Rate = 100.00M
-            //      },
-            //      new Room {
-            //        Id =  5,
-            //        Name = "Studio 5",
-            //        Description = "Measuring over 800 sqft and can accommodate up to 16 dancers comfortably. Our dance floor rests on a system of high-density rubber underlays and laminate wood flooring. It is fully air-conditioned and equipped with barres, mirrors and sound system compatible for audio input jack, USB device and SD card. This studio overlooks the street with many windows for natural lighting and ventilation.",
-            //        Image = "Studio5.jpg",
-            //        Size = "800 * 750",
-            //        Style= "K-pop;Hip Hop;Jazz",
-            //        Rate = 100.00M
-            //      }
-            //};
-            //roomViewModel.Rooms = roomList;
-            //roomViewModel.Error = 0;
-
             var result = _miscellaneousSerice.GetAllRoomType();
 
             if (result.Error != 0)
@@ -204,29 +148,52 @@ namespace StudioReservation.BackOffice.Controllers
             result.Room.StyleArr = result.Room.Style.Split(',');
             result.Room.RateOri = result.Room.Rate;
 
-            return View(result);
+            return View(result.Room);
         }
         [HttpPost]
         public ActionResult Edit(Room room,HttpPostedFileBase file)
         {
+            if (!ModelState.IsValid)
+            {
+                List<SelectListItem> styleOption = new List<SelectListItem>()
+                {
+                    new SelectListItem(){ Text = "Hip Hop", Value = "Hip Hop"},
+                    new SelectListItem(){ Text = "Jazz", Value = "Jazz"},
+                    new SelectListItem(){ Text = "K-pop", Value = "K-pop"},
+                    new SelectListItem(){ Text = "Latin", Value = "Latin"},
+                    new SelectListItem(){ Text = "Breaking", Value = "Breaking"}
+                };
+                ViewBag.Style = new MultiSelectList(styleOption, "Value", "Text");
+                room.RateOri = room.Rate;
+                return View(room);
+            }
 
+            string fileName = "";
+            string path = "";
+            bool newupload = false;
             if (file != null && file.ContentLength > 0)
             {
                 string fileExtension = Path.GetExtension(file.FileName);
                 string[] allowedExtensions = { ".pdf", ".jpg", ".png", ".gif" };
-
                 if (!allowedExtensions.Contains(fileExtension))
                 {
-
                     ViewBag.errorMessage = "Only PDF, JPG, PNG and GIF file extension supported.";
                     return View("Error");
                 }
-
-                //Store to ~/App_Data/Uploads folder
-                var fileName = Path.GetFileName(file.FileName);
-                var path = Path.Combine(Server.MapPath("~/Images/UserUploads"), fileName);
+                //Store to ~/Images/Uploads folder
+                fileName = DateTime.Now.ToString("yyyyMMddHHmmssfff")+ fileExtension;
+                path = Path.Combine(Server.MapPath("~/Images/UserUploads"), fileName);
                 file.SaveAs(path);
-                
+                newupload = true;
+            }
+            else if (file == null && !string.IsNullOrEmpty(room.Image))
+            {
+                fileName = room.Image;
+            }
+            else
+            {
+                return View("Error");
+            }         
                 //insert to db
                 room.Image = fileName;
                 room.Style = (room.StyleArr == null) ? "" : string.Join(",", room.StyleArr);
@@ -237,19 +204,14 @@ namespace StudioReservation.BackOffice.Controllers
 
                 if (result != 0)
                 {
-                    System.IO.File.Delete(path);
+                    if(newupload)
+                        System.IO.File.Delete(path);
 
                     ViewBag.ErrorCode = result.ToString();
                     return View("Error");
                 }
 
                 return RedirectToAction("Index");
-            }
-            else
-            {
-                ViewBag.errorMessage = "Something wrong with uploaded file.";
-                return View("Error");
-            }
-        }
+        }     
     }
 }
